\name{cpichg}
\docType{data}
\alias{cpichg}
\title{Data: Quarterly logarithmic change in percentage points of the Consumer Price Index for All Urban Consumers: All Items (CPIAUCSL) from 1947-04-01 to 2016-05-01}
\description{
Quarterly logarithmic change in percentage points of the Consumer Price Index for All Urban Consumers: All Items (CPIAUCSL) from 1947-04-01 to 2016-05-01 available at \url{https://fred.stlouisfed.org/series/CPIAUCSL}.
}
\usage{data("cpichg")}
\format{A \link{xts} object containing 276 observations from 1947-04-01 to 2016-05-01.}
\references{
US. Bureau of Labor Statistics, Consumer Price Index for All Urban Consumers: All Items [CPIAUCSL], retrieved from FRED, Federal Reserve Bank of St. Louis; https://fred.stlouisfed.org/series/CPIAUCSL, June 24, 2016.
}
\keyword{datasets}
