\name{dji30ret}
\docType{data}
\alias{dji30ret}
\title{data: Dow Jones 30 Constituents Closing Value Log Return}
\description{
This dataset is copied from the rugarch package of Ghalanos (2015).\cr
%

From the rugarch package:

Dow Jones 30 Constituents closing value log returns from 1987-03-16 to
2009-02-03 from Yahoo Finance. Note that AIG was replaced by KFT (Kraft Foods)
on September 22, 2008. This is not reflected in this data set as that would
bring the starting date of the data to 2001.
}
\usage{data(dji30ret)}
\format{A data.frame containing 5521x30 observations.}
\source{Yahoo Finance}
\references{Alexios Ghalanos (2015). rugarch: Univariate GARCH models. R package version 1.3-6.}
\keyword{datasets}
