% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethods.R
\docType{class}
\name{FitUGAS-class}
\alias{FitUGAS-class}
\title{FitUGAS Class}
\description{
Class for estimated models belonging to the Generalized Autoregressive Score family
}
\section{Slots}{

\describe{
\item{\code{model}}{Contains all the model information}

\item{\code{data}}{contains all the data in input and output}

\item{\code{fit}}{contains all the information about fit such as estimates}
}}

