\name{Goals}
\docType{data}
\alias{Goals}
\title{data: Goals scored by England against Scotland in international football matches.}
\description{

Number of goals scored by England against Scotland in international football matches.
This is a 116 x 2 \link{zoo} object spanning the period 1872-1987 with a yearly frequency.
The first column reports the number of goals scored by England against Scotland. The second
column is a dummy variable equal 1 for matches played in England.
This data set is taken from the Harvey (1989) pg 524.\cr

}
\usage{data("Goals")}
\format{A \link{zoo} object containing 116 x 2 observations.}
\source{Harvey (1989) pg. 524}
\references{
Harvey, A. C. (1990). Forecasting, structural time series models and the Kalman filter. Cambridge university press.
\url{https://cran.r-project.org/package=rugarch}.}
\keyword{datasets}
