\name{sp500ret}
\docType{data}
\alias{sp500ret}
\title{data: Daily logarithmic returns in percentage points of the S&P500 index from 1950-01-04 to 2016-06-24}
\description{
data: Daily logarithmic returns in percentage points of the  S&P500 index from 1950-01-04 to 2016-06-24 obtained from yahoo finance \url{http://finance.yahoo.com/q?s=^GSPC}.
}
\usage{data(sp500ret)}
\format{A \link{xts} object of dimension 16727 x 1 containing the daily logarithmic returns in percentage points from 1950-01-04 to 2016-06-24.}
\references{
\url{http://finance.yahoo.com/quote/^GSPC?ltr=1}
}
\keyword{datasets}
