\name{sp500rv}
\docType{data}
\alias{sp500rv}
\title{Data: SP500 Daily 5 minutes Realized Volatility from 2000-01-03 to 2000-01-10}
\description{
Oxford-Man Institute Daily 5 minutes Realized Volatility from 2000-01-03 to 2000-01-10 
for the SP500 Index available at \url{http://realized.oxford-man.ox.ac.uk/data}.
}
\usage{data("sp500rv")}
\format{A \link{xts} object containing 4,310 observations from 2000-01-03 to 2000-01-10.}
\references{
\url{http://realized.oxford-man.ox.ac.uk/data}
}
\keyword{datasets}
