% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GASspec.R
\name{specugas}
\alias{specugas}
\title{specugas}
\usage{
specugas(Dist = "norm")
}
\arguments{
\item{Dist}{The conditional distribution assumed for the data}
}
\value{
An SpecUGAS object
}
\description{
Univariate Generalised Autoregressive Score models specifications
}

