\name{StockIndexs}
\docType{data}
\alias{StockIndexs}
\title{data: Daily logarithmic returns in percentage points of the DAX, FTSEMIB and CAC40 from 2007-01-03 to 2016-06-24}
\description{
data: Daily logarithmic returns in percentage points of the DAX, FTSEMIB and CAC40 from 2007-01-03 to 2016-06-24 obtained from yahoo finance \url{https://finance.yahoo.com/}.
}
\usage{data(StockIndexs)}
\format{A \code{matrix} object of dimension 3 x 2445 containing the daily logarithmic returns in percentage points from 2007-01-03 to 2016-06-24. Missing values are simply removed.}
\references{
\url{https://finance.yahoo.com/}
}
\keyword{datasets}
