\name{StockIndices}
\docType{data}
\alias{StockIndices}
\title{Data: Daily logarithmic returns in percentage points of the DAX, FTSEMIB
and CAC40 from 2007-01-03 to 2016-06-24}
\description{
Daily logarithmic returns in percentage points of the DAX, FTSEMIB and CAC40
from 2007-01-03 to 2016-06-24 obtained from Yahoo.
}
\usage{data("StockIndices")}
\format{A \code{matrix} object of dimension 2,445 x 3 containing the daily logarithmic returns
in percentage points from 2007-01-03 to 2016-06-24. Missing values are simply removed.}
\references{
Yahoo finance.
}
\keyword{datasets}

