\name{usunp}
\docType{data}
\alias{usunp}
\title{US Monthly Civilian Unemployment Rate (UNRATE) from 1948-01-01 to 2016-05-01}
\description{
From  \url{https://fred.stlouisfed.org/series/UNRATE}:
The unemployment rate represents the number of unemployed as a percentage of the labor force. 
Labor force data are restricted to people 16 years of age and older, who currently reside in 1 of the 50 states or 
the District of Columbia, who do not reside in institutions (e.g., penal and mental facilities, 
homes for the aged), and who are not on active duty in the Armed Forces.
}
\usage{data("usunp")}
\format{A \link{xts} object containing 821 observations from 1948-01-01 to 2016-05-01.}
\references{
US. Bureau of Labor Statistics, Civilian Unemployment Rate [UNRATE], 
retrieved from FRED, Federal Reserve Bank of St. Louis; 
\url{https://fred.stlouisfed.org/series/UNRATE}, July 2, 2016.
}
\keyword{datasets}
