\name{GAS-package}
\alias{GAS-package}
\alias{GAS}
\docType{package}
\title{
Generalised Autoregressive Score models in R
}
\description{
The GAS package allows to simulate, estimate and forecast using univariate and multivariate Generalised Autoregressive Score (GAS) models (also known as Dynamic Conditional Score (DCS) models), see e.g. Creal et. al. (2013) and Harvey (2013).
}
\details{
\tabular{ll}{
Package: \tab GAS\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-06-06\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Leopoldo Catania, Kris Boudt and David Ardia

Maintainer: Leopoldo Catania <leopoldo.catania@uniroma2.it>
}
\references{
Creal, D., Koopman, S. J., & Lucas, A. (2013). Generalized autoregressive score models with applications. Journal of Applied Econometrics, 28(5), 777-795.\cr
%

Harvey, A. C. (2013). Dynamic models for volatility and heavy tails: with applications to financial and economic time series (Vol. 52). Cambridge University Press.\cr
}
\keyword{ package }
\seealso{
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
