\name{sp500ret}
\docType{data}
\alias{sp500ret}
\title{Data: Daily logarithmic returns in percentage points of the S&P500 index from 1950-01-04 to 2016-06-24}
\description{
Daily logarithmic returns in percentage points of the  S&P500 index from 1950-01-04 to 2016-06-24 
obtained from yahoo finance \url{http://finance.yahoo.com/quote/\%5EGSPC?ltr=1}.
}
\usage{data("sp500ret")}
\format{A \link{xts} object of dimension 16,727 x 1 containing the daily logarithmic 
returns in percentage points from 1950-01-04 to 2016-06-24.}
\source{Yahoo Finance}
\keyword{datasets}
